chore: rename `abi_command` module to `abi`
